Update TableVisual.cs
Added event handlers for:
- Dropped Cards
- Stacked Cards (placeholder)
- Took Cards (placeholder)
Made player count a global private variable to be used elsewhere.
Added function for dropping cards onto the table.
Added functions for checking and resizing the table cards when certain card amount thresholds are crossed.
Changed Deal To Table to only occur on the first round.
Changed the Turn Start event to only be raised by the master client instead of everyone.